Restructure run_loop.py for multi-broker per-user decision cycles
Fourth phase of the Alpaca integration, the big structural one: a user's
instrument list can now span both OANDA (forex) and Alpaca (equities/
crypto) in the same cycle. Previously account state/positions/exposure/
ExecutionService were fetched once per user per cycle, assuming a single
broker; now each broker kind present in a user's instrument list gets its
own BrokerCycleContext (own account, positions, exposure buckets,
execution service), built once and reused for every instrument routed to
that broker - same optimization as before, just keyed by broker instead
of assumed singular.

Also: widened risk_decisions.size_units and orders_fills.units from
Integer to Float (live Postgres ALTER + schema) and removed governor.py's
int() truncation for crypto sizing - a whole BTC/ETH costs tens of
thousands of dollars, so bounded-risk position sizing genuinely needs
fractional units, which the previous int() cast would have always
rounded to zero. Added dedicated equity/crypto correlation exposure
buckets so a long AAPL position no longer shares a correlation cap with
FX pairs. FX-only session scoring (pair_session_score,
compute_prior_session_range) is now skipped for non-forex instruments
rather than silently misapplied.

Verified live end-to-end: a real cycle against the admin's actual
credentials, instrument list spanning EUR_USD (OANDA) + AAPL + BTC/USD
(Alpaca), completed with zero exceptions - correct per-broker account/
position isolation, correct exposure bucket routing. Alpaca price models
correctly declined to train yet (this environment's simulated market data
currently has ~64 M15 bars for AAPL, below the 100-bar minimum) - a real
data-availability characteristic that resolves as more real bars
accumulate, not a code defect.

+++ src/data/db.py
@@ -286,7 +286,9 @@ risk_decisions = Table(
     Column("time", DateTime(timezone=True), nullable=False, index=True),
     Column("approved", Boolean, nullable=False),
     Column("reason", String, nullable=False),
-    Column("size_units", Integer, nullable=True),
+    # Float for the same reason as orders_fills.units above — crypto needs
+    # fractional sizing.
+    Column("size_units", Float, nullable=True),
     Column("gates_json", Text, nullable=True),
 )
 
@@ -301,7 +303,13 @@ orders_fills = Table(
     Column("broker_order_id", String, nullable=True),
     Column("broker_transaction_id", String, nullable=True),
     Column("instrument", String, nullable=False),
-    Column("units", Integer, nullable=False),
+    # Float, not Integer: crypto (Alpaca) sizes in fractional units — a
+    # whole BTC/ETH costs tens of thousands of dollars, so risk-bounded
+    # sizing genuinely needs fractional amounts. Forex/equity units are
+    # still always whole numbers in practice; Float represents those
+    # exactly too, so this is a strict widening, not a behavior change for
+    # existing OANDA rows.
+    Column("units", Float, nullable=False),
     Column("status", String, nullable=False),
     Column("fill_price", Float, nullable=True),
     Column("spread", Float, nullable=True),
